add the baseic function of game engine, create game , join game

+++ resources/game.go
@@ -2,7 +2,7 @@ package resources
 
 //Game List of Games running by the game engin
 type Game struct {
-	ID         int
+	ID         int64
 	IsPublic   bool
 	User       UserModel
 	TimeLine   []string
@@ -13,5 +13,5 @@ type Game struct {
 //CreateGameModel to create new game
 type CreateGameModel struct {
 	IsPublic bool
-	UserID   int
+	UserID   uint64
 }
